Enhance data handling and processing in MSIConverter
- Updated data types from float32 to float64 for improved precision in lightweight and spatial data converters.
- Introduced Total Ion Current (TIC) image generation for both 2D slices and 3D datasets in SpatialDataConverter.
- Added functionality to calculate and store average mass spectrum using only non-empty pixels.
- Implemented pixel count tracking for better metadata representation.
- Enhanced error handling and logging during data processing.
- Created a new napari viewer script for visualizing rasterized MSI data.
- Added comprehensive documentation for accessing and manipulating SpatialData objects, including TIC images and average mass spectra.

+++ msiconvert/converters/lightweight_converter.py
@@ -103,7 +103,7 @@ class LightweightConverter(BaseMSIConverter):
         self.root.create_dataset(
             'sparse_data/data',
             shape=(0,),
-            dtype=np.float32,
+            dtype=np.float64,
             compressor=compressor,
             chunks=(10000,)
         )
@@ -182,7 +182,7 @@ class LightweightConverter(BaseMSIConverter):
         indices_array.resize((new_size, 2))
         
         # Convert data to numpy arrays for efficient bulk insertion
-        data_np = np.array(data_buffer, dtype=np.float32)
+        data_np = np.array(data_buffer, dtype=np.float64)
         indices_np = np.array(indices_buffer, dtype=np.int32)
         
         # Store data in a single operation
